Apply changes from upstream
https://github.com/KDE/konsole/commit/f65e5a5797bb5d633ef7339ff5f285e918fda401
https://github.com/KDE/konsole/commit/be7181c93ad0f36953555e2cee402e193da2d5a9
https://github.com/KDE/konsole/commit/5128781a824c26dc2746650fea0ae9f95861b9d8
https://github.com/KDE/konsole/commit/f24aa818f97b6bebdcc4ce62c927994232b6f282
https://github.com/KDE/konsole/commit/674de973f9d2c85b49929f885d0cfb135fa6bbab

+++ src/History.h
@@ -282,7 +282,7 @@ protected:
     CompactHistoryBlockList& _blockListRef;
     CharacterFormat* _formatArray;
     quint16 _length;
-    quint16* _text;
+    uint*   _text;
     quint16 _formatLength;
     bool _wrapped;
 };
